+++ src/widget/text_widget.rs
@@ -17,6 +17,7 @@ use opengl_graphics::GlGraphics;
 use piston_window::*;
 
 use crate::core::point::*;
+use crate::core::callbacks::*;
 use crate::widget::config::*;
 use crate::widget::widget::*;
 
@@ -24,6 +25,7 @@ use crate::widget::widget::*;
 /// no accessable objects, they are all internal to `TextWidget`'s implementation.
 pub struct TextWidget {
     config: Configurable,
+    callbacks: CallbackStore,
     font_cache: Glyphs,
     text: String,
     font_size: u32,
@@ -44,6 +46,7 @@ impl TextWidget {
 
         Self {
             config: Configurable::new(),
+            callbacks: CallbackStore::new(),
             font_cache: glyphs,
             text,
             font_size,
@@ -135,6 +138,10 @@ impl TextWidget {
 impl Widget for TextWidget {
     fn config(&mut self) -> &mut Configurable {
         &mut self.config
+    }
+
+    fn callbacks(&mut self) -> &mut CallbackStore {
+        &mut self.callbacks
     }
 
     /// Draws the contents of the widget.